+++ build_tools/aws-sdk-code-generator/lib/aws-sdk-code-generator/plugin_list.rb
@@ -46,6 +46,7 @@ module AwsSdkCodeGenerator
         'Aws::Plugins::RetryErrors' => "#{core_plugins}/retry_errors.rb",
         'Aws::Plugins::GlobalConfiguration' => "#{core_plugins}/global_configuration.rb",
         'Aws::Plugins::RegionalEndpoint' => "#{core_plugins}/regional_endpoint.rb",
+        'Aws::Plugins::EndpointDiscovery' => "#{core_plugins}/endpoint_discovery.rb",
         'Aws::Plugins::ResponsePaging' => "#{core_plugins}/response_paging.rb",
         'Aws::Plugins::StubResponses' => "#{core_plugins}/stub_responses.rb",
         'Aws::Plugins::IdempotencyToken' => "#{core_plugins}/idempotency_token.rb",